test: split mitm connection hook tests
