--- crates/runner/mitm-addon/tests/test_connection_hooks.py
@@ -1,1181 +1,0 @@
-"""Tests for mitm addon connection-level hooks."""
-
-import json
-import threading
-import time
-from collections.abc import Callable
-from pathlib import Path
-from unittest.mock import MagicMock, patch
-
-import pytest
-from mitmproxy import tcp
-from mitmproxy.flow import Error
-
-import flow_metadata_keys as metadata_keys
-import logging_utils
-import mitm_addon
-import registry
-import usage
-from tests.jsonl_log_helpers import read_jsonl_entries_after_flush
-from tests.pending_helpers import assert_pending
-from tests.timestamp_helpers import assert_utc_millisecond_timestamp
-from tests.usage_buffer_helpers import RecordingEnqueue, event
-from tests.usage_helpers import install_recording_usage_timer
-
-
-def wait_for_usage_flush_worker_to_stop(timeout: float = 1.0) -> None:
-    acquired = mitm_addon._usage_flush_signal_lock.acquire(timeout=timeout)
-    assert acquired
-    mitm_addon._usage_flush_signal_lock.release()
-
-
-def reset_runner_usage_flush_state() -> None:
-    mitm_addon._usage_flush_requested.clear()
-    mitm_addon._last_jsonl_flush_request_id = None
-    wait_for_usage_flush_worker_to_stop()
-
-
-class _InstrumentedFlushOwnerLock:
-    def __init__(self) -> None:
-        self._lock = threading.Lock()
-        self.blocking_acquire_started = threading.Event()
-
-    def acquire(self, blocking: bool = True) -> bool:
-        if blocking:
-            self.blocking_acquire_started.set()
-        return self._lock.acquire(blocking)
-
-    def release(self) -> None:
-        self._lock.release()
-
-
-_ScheduledTcpDrain = tuple[Callable[[tcp.TCPFlow], None], tcp.TCPFlow]
-
-
-def _capture_tcp_drains(monkeypatch: pytest.MonkeyPatch) -> list[_ScheduledTcpDrain]:
-    scheduled: list[_ScheduledTcpDrain] = []
-
-    def call_soon(callback: Callable[[tcp.TCPFlow], None], flow: tcp.TCPFlow) -> None:
-        scheduled.append((callback, flow))
-
-    monkeypatch.setattr(mitm_addon, "_call_soon", call_soon)
-    return scheduled
-
-
-def _run_scheduled_tcp_drains(scheduled: list[_ScheduledTcpDrain]) -> None:
-    pending = list(scheduled)
-    scheduled.clear()
-    for callback, flow in pending:
-        callback(flow)
-
-
-class TestDoneHook:
-    """Tests for the done() graceful shutdown hook."""
-
-    def test_done_shuts_down_executor(self):
-        """done() should call shutdown(wait=True) on the executor."""
-        mock_executor = MagicMock()
-        with (
-            patch.object(usage, "flush_usage_events") as flush_usage_events,
-            patch.object(usage.webhook, "usage_executor", mock_executor),
-            patch.object(
-                mitm_addon.auth_base_forwarder,
-                "shutdown_forward_request_executor",
-            ) as shutdown_forward_request_executor,
-            patch.object(mitm_addon, "shutdown_log_writer") as shutdown_log_writer,
-        ):
-            mitm_addon.done()
-        flush_usage_events.assert_called_once_with(trigger="shutdown")
-        # concurrent.futures boundary: done() must gracefully shut down the pool (#9991).
-        mock_executor.shutdown.assert_called_once_with(wait=True)
-        shutdown_forward_request_executor.assert_called_once_with(wait=False)
-        shutdown_log_writer.assert_called_once_with()
-
-    def test_done_waits_for_runner_flush_before_executor_shutdown(self):
-        """done() must not shut down the executor while a SIGUSR1 flush is enqueueing."""
-
-        class _InstrumentedLock:
-            def __init__(self) -> None:
-                self._lock = threading.Lock()
-                self.blocking_acquire_started = threading.Event()
-
-            def acquire(self, blocking: bool = True) -> bool:
-                if blocking:
-                    self.blocking_acquire_started.set()
-                return self._lock.acquire(blocking)
-
-            def release(self) -> None:
-                self._lock.release()
-
-            def __enter__(self):
-                self.acquire()
-                return self
-
-            def __exit__(self, exc_type, exc, traceback) -> None:
-                del exc_type, exc, traceback
-                self.release()
-
-        lock = _InstrumentedLock()
-        runner_flush_started = threading.Event()
-        release_runner_flush = threading.Event()
-        shutdown_called = threading.Event()
-        calls: list[str] = []
-
-        def flush_usage_events(*, trigger: str) -> int:
-            calls.append(f"flush:{trigger}")
-            if trigger == "runner":
-                runner_flush_started.set()
-                if not release_runner_flush.wait(timeout=1):
-                    calls.append("runner_flush_timeout")
-            return 0
-
-        def shutdown(*, wait: bool) -> None:
-            calls.append(f"shutdown:{wait}")
-            shutdown_called.set()
-
-        mock_executor = MagicMock()
-        mock_executor.shutdown.side_effect = shutdown
-
-        with (
-            patch.object(mitm_addon, "_usage_flush_signal_lock", lock),
-            patch.object(usage, "flush_usage_events", side_effect=flush_usage_events),
-            patch.object(usage.webhook, "usage_executor", mock_executor),
-            patch.object(
-                mitm_addon.auth_base_forwarder,
-                "shutdown_forward_request_executor",
-                lambda *, wait: calls.append(f"auth-base:shutdown:{wait}"),
-            ),
-            patch.object(mitm_addon, "shutdown_log_writer", lambda: calls.append("jsonl:shutdown")),
-        ):
-            mitm_addon._handle_runner_usage_flush_signal(0, None)
-            assert runner_flush_started.wait(timeout=1)
-
-            done_thread = threading.Thread(target=mitm_addon.done)
-            done_thread.start()
-            assert lock.blocking_acquire_started.wait(timeout=1)
-            assert not shutdown_called.is_set()
-
-            release_runner_flush.set()
-            done_thread.join(timeout=1)
-
-        assert not done_thread.is_alive()
-        assert calls == [
-            "flush:runner",
-            "flush:shutdown",
-            "shutdown:True",
-            "auth-base:shutdown:False",
-            "jsonl:shutdown",
-        ]
-
-    def test_done_shuts_down_executor_when_flush_fails(self):
-        mock_executor = MagicMock()
-
-        with (
-            patch.object(
-                usage,
-                "flush_usage_events",
-                side_effect=RuntimeError("flush failed"),
-            ) as flush_usage_events,
-            patch.object(usage.webhook, "usage_executor", mock_executor),
-            patch.object(
-                mitm_addon.auth_base_forwarder,
-                "shutdown_forward_request_executor",
-            ) as shutdown_forward_request_executor,
-            patch.object(mitm_addon, "shutdown_log_writer") as shutdown_log_writer,
-            pytest.raises(RuntimeError, match="flush failed"),
-        ):
-            mitm_addon.done()
-
-        flush_usage_events.assert_called_once_with(trigger="shutdown")
-        mock_executor.shutdown.assert_called_once_with(wait=True)
-        shutdown_forward_request_executor.assert_called_once_with(wait=False)
-        shutdown_log_writer.assert_called_once_with()
-
-
-class TestRunnerUsageFlushSignal:
-    """Tests for runner-triggered usage buffer flush requests."""
-
-    def test_signal_handler_flushes_usage_in_background(self, tmp_path):
-        reset_runner_usage_flush_state()
-        flushed = threading.Event()
-        snapshotted = threading.Event()
-        pending_path = tmp_path / "usage-pending"
-        request_path = tmp_path / "usage-flush-request"
-        usage.set_pending_path(str(pending_path), usage_state_id="runner-state")
-        request_path.write_text(
-            json.dumps(
-                {
-                    "usageStateId": "runner-state",
-                    "flushRequestId": "request-1",
-                    "requestedAtMs": 1_770_000_000_000,
-                }
-            )
-        )
-
-        def flush_usage_events(*, trigger: str) -> int:
-            assert trigger == "runner"
-            usage.counters.increment_pending_reports()
-            usage.counters.decrement_pending_reports()
-            flushed.set()
-            return 0
-
-        original_write_pending_snapshot = usage.write_pending_snapshot
-
-        def write_pending_snapshot(*, flush_request_id: str | None = None) -> None:
-            original_write_pending_snapshot(flush_request_id=flush_request_id)
-            snapshotted.set()
-
-        try:
-            with (
-                patch.object(usage, "flush_usage_events", side_effect=flush_usage_events),
-                patch.object(
-                    usage,
-                    "write_pending_snapshot",
-                    side_effect=write_pending_snapshot,
-                ),
-            ):
-                mitm_addon._handle_runner_usage_flush_signal(0, None)
-                assert flushed.wait(timeout=1)
-                assert snapshotted.wait(timeout=1)
-                wait_for_usage_flush_worker_to_stop()
-
-            assert_pending(
-                pending_path,
-                flows=0,
-                buffered=0,
-                reports=0,
-                flush_request_id="request-1",
-            )
-        finally:
-            usage.set_pending_path("")
-
-    def test_signal_handler_writes_snapshot_when_flush_fails(self, tmp_path):
-        reset_runner_usage_flush_state()
-        snapshotted = threading.Event()
-        pending_path = tmp_path / "usage-pending"
-        request_path = tmp_path / "usage-flush-request"
-        usage.set_pending_path(str(pending_path), usage_state_id="runner-state")
-        usage.counters.increment_pending_reports()
-        request_path.write_text(
-            json.dumps(
-                {
-                    "usageStateId": "runner-state",
-                    "flushRequestId": "request-1",
-                    "requestedAtMs": 1_770_000_000_000,
-                }
-            )
-        )
-
-        original_write_pending_snapshot = usage.write_pending_snapshot
-
-        def write_pending_snapshot(*, flush_request_id: str | None = None) -> None:
-            original_write_pending_snapshot(flush_request_id=flush_request_id)
-            snapshotted.set()
-
-        try:
-            with (
-                patch.object(
-                    usage,
-                    "flush_usage_events",
-                    side_effect=RuntimeError("flush failed"),
-                ),
-                patch.object(
-                    usage,
-                    "write_pending_snapshot",
-                    side_effect=write_pending_snapshot,
-                ),
-                patch.object(mitm_addon.ctx, "log", MagicMock(), create=True),
-            ):
-                mitm_addon._handle_runner_usage_flush_signal(0, None)
-                assert snapshotted.wait(timeout=1)
-                wait_for_usage_flush_worker_to_stop()
-
-            assert_pending(
-                pending_path,
-                flows=0,
-                buffered=0,
-                reports=1,
-                flush_request_id="request-1",
-            )
-        finally:
-            usage.counters.decrement_pending_reports()
-            usage.set_pending_path("")
-
-    def test_signal_handler_acknowledges_jsonl_flush_request(self, tmp_path):
-        reset_runner_usage_flush_state()
-        pending_path = tmp_path / "usage-pending"
-        request_path = tmp_path / "jsonl-flush-request"
-        state_path = tmp_path / "jsonl-flush-state"
-        log_path = tmp_path / "network.jsonl"
-        usage.set_pending_path(str(pending_path), usage_state_id="runner-state")
-        request_path.write_text(
-            json.dumps(
-                {
-                    "usageStateId": "runner-state",
-                    "flushRequestId": "jsonl-request-1",
-                    "requestedAtMs": 1_770_000_000_000,
-                    "path": str(log_path),
-                }
-            )
-        )
-
-        try:
-            with (
-                patch.object(mitm_addon, "__file__", str(tmp_path / "mitm_addon.py")),
-                patch.object(logging_utils.ctx, "log", MagicMock(), create=True),
-            ):
-                logging_utils.log_network_entry(str(log_path), {"action": "ALLOW"})
-                mitm_addon._handle_runner_usage_flush_signal(0, None)
-                wait_for_usage_flush_worker_to_stop()
-
-            entry = json.loads(log_path.read_text().strip())
-            assert entry["action"] == "ALLOW"
-            state = json.loads(state_path.read_text())
-            assert state == {
-                "pid": state["pid"],
-                "usageStateId": "runner-state",
-                "updatedAtMs": state["updatedAtMs"],
-                "flushRequestId": "jsonl-request-1",
-                "path": str(log_path),
-                "pending": 0,
-            }
-        finally:
-            usage.set_pending_path("")
-
-    def test_jsonl_flush_request_rejects_unsafe_request_id(self, tmp_path):
-        reset_runner_usage_flush_state()
-        pending_path = tmp_path / "usage-pending"
-        request_path = tmp_path / "jsonl-flush-request"
-        state_path = tmp_path / "jsonl-flush-state"
-        log_path = tmp_path / "network.jsonl"
-        usage.set_pending_path(str(pending_path), usage_state_id="runner-state")
-        request_path.write_text(
-            json.dumps(
-                {
-                    "usageStateId": "runner-state",
-                    "flushRequestId": "../jsonl-request-1",
-                    "requestedAtMs": 1_770_000_000_000,
-                    "path": str(log_path),
-                }
-            )
-        )
-
-        try:
-            with (
-                patch.object(mitm_addon, "__file__", str(tmp_path / "mitm_addon.py")),
-                patch.object(mitm_addon, "flush_log_path") as flush_log_path,
-            ):
-                mitm_addon._handle_runner_usage_flush_signal(0, None)
-                wait_for_usage_flush_worker_to_stop()
-
-            flush_log_path.assert_not_called()
-            assert not state_path.exists()
-        finally:
-            usage.set_pending_path("")
-
-    def test_jsonl_flush_failure_writes_pending_state(self, tmp_path):
-        reset_runner_usage_flush_state()
-        pending_path = tmp_path / "usage-pending"
-        request_path = tmp_path / "jsonl-flush-request"
-        state_path = tmp_path / "jsonl-flush-state"
-        log_path = tmp_path / "network.jsonl"
-        log = MagicMock()
-        usage.set_pending_path(str(pending_path), usage_state_id="runner-state")
-        request_path.write_text(
-            json.dumps(
-                {
-                    "usageStateId": "runner-state",
-                    "flushRequestId": "jsonl-request-1",
-                    "requestedAtMs": 1_770_000_000_000,
-                    "path": str(log_path),
-                }
-            )
-        )
-
-        try:
-            with (
-                patch.object(mitm_addon, "__file__", str(tmp_path / "mitm_addon.py")),
-                patch.object(mitm_addon, "flush_log_path", side_effect=RuntimeError("secret")),
-                patch.object(mitm_addon.ctx, "log", log, create=True),
-            ):
-                mitm_addon._handle_runner_usage_flush_signal(0, None)
-                wait_for_usage_flush_worker_to_stop()
-
-            state = json.loads(state_path.read_text())
-            assert state == {
-                "pid": state["pid"],
-                "usageStateId": "runner-state",
-                "updatedAtMs": state["updatedAtMs"],
-                "flushRequestId": "jsonl-request-1",
-                "path": str(log_path),
-                "pending": 1,
-            }
-            log.warn.assert_called_once()
-            warning = log.warn.call_args.args[0]
-            assert "RuntimeError" in warning
-            assert "secret" not in warning
-        finally:
-            usage.set_pending_path("")
-
-    def test_signal_handler_does_not_reprocess_acknowledged_jsonl_flush_request(self, tmp_path):
-        reset_runner_usage_flush_state()
-        pending_path = tmp_path / "usage-pending"
-        request_path = tmp_path / "jsonl-flush-request"
-        log_path = tmp_path / "network.jsonl"
-        usage.set_pending_path(str(pending_path), usage_state_id="runner-state")
-        request_path.write_text(
-            json.dumps(
-                {
-                    "usageStateId": "runner-state",
-                    "flushRequestId": "jsonl-request-1",
-                    "requestedAtMs": 1_770_000_000_000,
-                    "path": str(log_path),
-                }
-            )
-        )
-
-        try:
-            with (
-                patch.object(mitm_addon, "__file__", str(tmp_path / "mitm_addon.py")),
-                patch.object(mitm_addon, "flush_log_path") as flush_log_path,
-                patch.object(mitm_addon.ctx, "log", MagicMock(), create=True),
-            ):
-                mitm_addon._handle_runner_usage_flush_signal(0, None)
-                wait_for_usage_flush_worker_to_stop()
-                mitm_addon._handle_runner_usage_flush_signal(0, None)
-                wait_for_usage_flush_worker_to_stop()
-
-            flush_log_path.assert_called_once_with(str(log_path))
-        finally:
-            usage.set_pending_path("")
-
-    def test_jsonl_flush_failure_is_retryable(self, tmp_path):
-        reset_runner_usage_flush_state()
-        pending_path = tmp_path / "usage-pending"
-        request_path = tmp_path / "jsonl-flush-request"
-        state_path = tmp_path / "jsonl-flush-state"
-        log_path = tmp_path / "network.jsonl"
-        usage.set_pending_path(str(pending_path), usage_state_id="runner-state")
-        request_path.write_text(
-            json.dumps(
-                {
-                    "usageStateId": "runner-state",
-                    "flushRequestId": "jsonl-request-1",
-                    "requestedAtMs": 1_770_000_000_000,
-                    "path": str(log_path),
-                }
-            )
-        )
-
-        try:
-            with (
-                patch.object(mitm_addon, "__file__", str(tmp_path / "mitm_addon.py")),
-                patch.object(mitm_addon.ctx, "log", MagicMock(), create=True),
-            ):
-                with patch.object(
-                    mitm_addon,
-                    "flush_log_path",
-                    side_effect=RuntimeError("flush failed"),
-                ) as failed_flush:
-                    mitm_addon._handle_runner_usage_flush_signal(0, None)
-                    wait_for_usage_flush_worker_to_stop()
-
-                with patch.object(mitm_addon, "flush_log_path") as retry_flush:
-                    mitm_addon._handle_runner_usage_flush_signal(0, None)
-                    wait_for_usage_flush_worker_to_stop()
-
-            failed_flush.assert_called_once_with(str(log_path))
-            retry_flush.assert_called_once_with(str(log_path))
-            state = json.loads(state_path.read_text())
-            assert state["pending"] == 0
-        finally:
-            usage.set_pending_path("")
-
-    def test_runner_flush_failure_warns_without_error_text(self):
-        log = MagicMock()
-
-        with (
-            patch.object(mitm_addon.ctx, "log", log, create=True),
-            patch.object(
-                usage,
-                "flush_usage_events",
-                side_effect=RuntimeError("secret-token"),
-            ),
-        ):
-            mitm_addon._flush_usage_for_runner_request()
-
-        log.warn.assert_called_once()
-        message = log.warn.call_args.args[0]
-        assert "RuntimeError" in message
-        assert "secret-token" not in message
-
-    def test_runner_flush_failure_snapshot_includes_retryable_buffered_usage(self, tmp_path):
-        pending_path = tmp_path / "usage-pending"
-        request_path = tmp_path / "usage-flush-request"
-        usage.set_pending_path(str(pending_path), usage_state_id="runner-state")
-        request_path.write_text(
-            json.dumps(
-                {
-                    "usageStateId": "runner-state",
-                    "flushRequestId": "request-1",
-                    "requestedAtMs": 1_770_000_000_000,
-                }
-            )
-        )
-        enqueue_calls = 0
-
-        def fail_enqueue_webhook(
-            url: str,
-            sandbox_token: str,
-            payload: dict,
-            proxy_log_path: str,
-            log_type: str,
-            delivery_outcome_callback: Callable[[usage.webhook.WebhookDeliveryOutcome], None],
-        ) -> bool:
-            nonlocal enqueue_calls
-            del url, sandbox_token, payload, proxy_log_path, log_type, delivery_outcome_callback
-            enqueue_calls += 1
-            raise OSError("no threads")
-
-        usage.reset_usage_buffer_for_tests(enqueue_webhook=fail_enqueue_webhook)
-        usage.buffer_usage_events(
-            "https://api.test/api/webhooks/agent/usage-event",
-            "token-a",
-            "run-1",
-            [
-                {
-                    "idempotencyKey": "source-1",
-                    "kind": "model",
-                    "provider": "claude-sonnet-4-6",
-                    "category": "tokens.input",
-                    "quantity": 1,
-                }
-            ],
-            str(tmp_path / "proxy.jsonl"),
-        )
-
-        try:
-            with patch.object(mitm_addon.ctx, "log", MagicMock(), create=True):
-                mitm_addon._flush_usage_for_runner_request()
-
-            assert enqueue_calls == 1
-            assert_pending(
-                pending_path,
-                flows=0,
-                buffered=1,
-                reports=0,
-                flush_request_id="request-1",
-            )
-        finally:
-            usage.set_pending_path("")
-
-    def test_signal_waits_for_active_timer_flush_before_ack_snapshot(self, tmp_path):
-        reset_runner_usage_flush_state()
-        flush_owner_lock = _InstrumentedFlushOwnerLock()
-        timer_enqueue_started = threading.Event()
-        release_timer_enqueue = threading.Event()
-        enqueued_runs: list[str] = []
-        pending_path = tmp_path / "usage-pending"
-        request_path = tmp_path / "usage-flush-request"
-        proxy_log_path = str(tmp_path / "proxy.jsonl")
-        usage.set_pending_path(str(pending_path), usage_state_id="runner-state")
-        request_path.write_text(
-            json.dumps(
-                {
-                    "usageStateId": "runner-state",
-                    "flushRequestId": "request-1",
-                    "requestedAtMs": 1_770_000_000_000,
-                }
-            )
-        )
-
-        def enqueue_webhook(url, sandbox_token, payload, path, log_type):
-            del url, sandbox_token
-            assert log_type == "usage_event"
-            assert path == proxy_log_path
-            enqueued_runs.append(payload["runId"])
-            if payload["runId"] == "run-1":
-                timer_enqueue_started.set()
-                assert release_timer_enqueue.wait(timeout=2)
-
-        enqueue = RecordingEnqueue(side_effect=enqueue_webhook)
-        timers = install_recording_usage_timer(
-            enqueue_webhook=enqueue,
-            flush_owner_lock=flush_owner_lock,
-        )
-        usage.buffer_usage_events(
-            "https://api.test/api/webhooks/agent/usage-event",
-            "token-a",
-            "run-1",
-            [event(source_key="source-1")],
-            proxy_log_path,
-        )
-        timer_thread = threading.Thread(target=timers[0].callback)
-        timer_thread_started = False
-
-        try:
-            timer_thread.start()
-            timer_thread_started = True
-            assert timer_enqueue_started.wait(timeout=1)
-
-            mitm_addon._handle_runner_usage_flush_signal(0, None)
-            assert flush_owner_lock.blocking_acquire_started.wait(timeout=1)
-            state_before_release = json.loads(pending_path.read_text())
-            assert "flushRequestId" not in state_before_release
-
-            usage.buffer_usage_events(
-                "https://api.test/api/webhooks/agent/usage-event",
-                "token-a",
-                "run-2",
-                [event(source_key="source-2")],
-                proxy_log_path,
-            )
-            assert len(timers) == 2
-
-            release_timer_enqueue.set()
-            timer_thread.join(timeout=1)
-            wait_for_usage_flush_worker_to_stop()
-
-            assert not timer_thread.is_alive()
-            assert enqueued_runs == ["run-1", "run-2"]
-            assert timers[1].cancelled is True
-            assert_pending(
-                pending_path,
-                flows=0,
-                buffered=0,
-                reports=0,
-                flush_request_id="request-1",
-            )
-        finally:
-            release_timer_enqueue.set()
-            if timer_thread_started:
-                timer_thread.join(timeout=1)
-            wait_for_usage_flush_worker_to_stop()
-            usage.set_pending_path("")
-
-    def test_signal_retries_failed_active_timer_flush_before_ack_snapshot(self, tmp_path):
-        reset_runner_usage_flush_state()
-        flush_owner_lock = _InstrumentedFlushOwnerLock()
-        first_enqueue_started = threading.Event()
-        release_first_enqueue = threading.Event()
-        enqueued_run_ids: list[str] = []
-        enqueued_idempotency_keys: list[str] = []
-        timer_errors: list[str] = []
-        pending_path = tmp_path / "usage-pending"
-        request_path = tmp_path / "usage-flush-request"
-        proxy_log_path = str(tmp_path / "proxy.jsonl")
-        usage.set_pending_path(str(pending_path), usage_state_id="runner-state")
-        request_path.write_text(
-            json.dumps(
-                {
-                    "usageStateId": "runner-state",
-                    "flushRequestId": "request-1",
-                    "requestedAtMs": 1_770_000_000_000,
-                }
-            )
-        )
-
-        def enqueue_webhook(url, sandbox_token, payload, path, log_type):
-            del url, sandbox_token, path
-            assert log_type == "usage_event"
-            enqueued_run_ids.append(payload["runId"])
-            enqueued_idempotency_keys.append(payload["events"][0]["idempotencyKey"])
-            if len(enqueued_run_ids) == 1:
-                first_enqueue_started.set()
-                assert release_first_enqueue.wait(timeout=2)
-                raise OSError("timer failed")
-
-        enqueue = RecordingEnqueue(side_effect=enqueue_webhook)
-        timers = install_recording_usage_timer(
-            enqueue_webhook=enqueue,
-            flush_owner_lock=flush_owner_lock,
-        )
-        usage.buffer_usage_events(
-            "https://api.test/api/webhooks/agent/usage-event",
-            "token-a",
-            "run-1",
-            [event(source_key="source-1")],
-            proxy_log_path,
-        )
-
-        def timer_flush():
-            try:
-                timers[0].callback()
-            except OSError as exc:
-                timer_errors.append(str(exc))
-
-        timer_thread = threading.Thread(target=timer_flush)
-        timer_thread_started = False
-
-        try:
-            timer_thread.start()
-            timer_thread_started = True
-            assert first_enqueue_started.wait(timeout=1)
-
-            mitm_addon._handle_runner_usage_flush_signal(0, None)
-            assert flush_owner_lock.blocking_acquire_started.wait(timeout=1)
-            state_before_release = json.loads(pending_path.read_text())
-            assert "flushRequestId" not in state_before_release
-
-            release_first_enqueue.set()
-            timer_thread.join(timeout=1)
-            wait_for_usage_flush_worker_to_stop()
-
-            assert not timer_thread.is_alive()
-            assert timer_errors == ["timer failed"]
-            assert enqueued_run_ids == ["run-1", "run-1"]
-            assert enqueued_idempotency_keys[0] == enqueued_idempotency_keys[1]
-            assert len(timers) == 2
-            assert timers[0].cancelled is True
-            assert timers[1].cancelled is True
-            assert_pending(
-                pending_path,
-                flows=0,
-                buffered=0,
-                reports=0,
-                flush_request_id="request-1",
-            )
-        finally:
-            release_first_enqueue.set()
-            if timer_thread_started:
-                timer_thread.join(timeout=1)
-            wait_for_usage_flush_worker_to_stop()
-            usage.set_pending_path("")
-
-    def test_signal_during_active_flush_runs_follow_up_flush(self):
-        reset_runner_usage_flush_state()
-        first_flush_started = threading.Event()
-        release_first_flush = threading.Event()
-        second_flush_completed = threading.Event()
-        worker_timed_out = threading.Event()
-        flush_triggers: list[str] = []
-
-        def flush_usage_events(*, trigger: str) -> int:
-            flush_triggers.append(trigger)
-            if len(flush_triggers) == 1:
-                first_flush_started.set()
-                if not release_first_flush.wait(timeout=2):
-                    worker_timed_out.set()
-            else:
-                second_flush_completed.set()
-            return 0
-
-        with patch.object(usage, "flush_usage_events", side_effect=flush_usage_events):
-            mitm_addon._handle_runner_usage_flush_signal(0, None)
-            assert first_flush_started.wait(timeout=1)
-
-            mitm_addon._handle_runner_usage_flush_signal(0, None)
-            release_first_flush.set()
-
-            assert second_flush_completed.wait(timeout=1)
-            wait_for_usage_flush_worker_to_stop()
-
-        assert not worker_timed_out.is_set()
-        assert flush_triggers == ["runner", "runner"]
-
-    def test_failed_signal_flush_releases_worker_for_later_signal(self, mitm_ctx):
-        reset_runner_usage_flush_state()
-        second_flush_completed = threading.Event()
-        flush_triggers: list[str] = []
-
-        def flush_usage_events(*, trigger: str) -> int:
-            flush_triggers.append(trigger)
-            if len(flush_triggers) == 1:
-                raise RuntimeError("flush failed")
-            second_flush_completed.set()
-            return 0
-
-        with (
-            mitm_ctx() as log,
-            patch.object(usage, "flush_usage_events", side_effect=flush_usage_events),
-        ):
-            mitm_addon._handle_runner_usage_flush_signal(0, None)
-            wait_for_usage_flush_worker_to_stop()
-
-            mitm_addon._handle_runner_usage_flush_signal(0, None)
-            assert second_flush_completed.wait(timeout=1)
-            wait_for_usage_flush_worker_to_stop()
-
-        log.warn.assert_called_once()
-        assert flush_triggers == ["runner", "runner"]
-
-    def test_signal_retryable_delivery_failure_recovers_on_later_signal(
-        self,
-        tmp_path,
-        sync_usage_executor,
-        usage_webhook_server,
-    ):
-        del sync_usage_executor
-        reset_runner_usage_flush_state()
-        pending_path = tmp_path / "usage-pending"
-        request_path = tmp_path / "usage-flush-request"
-        proxy_log_path = str(tmp_path / "proxy.jsonl")
-        usage.set_pending_path(str(pending_path), usage_state_id="runner-state")
-        request_path.write_text(
-            json.dumps(
-                {
-                    "usageStateId": "runner-state",
-                    "flushRequestId": "request-1",
-                    "requestedAtMs": 1_770_000_000_000,
-                }
-            )
-        )
-        usage.buffer_usage_events(
-            usage_webhook_server.url("/usage"),
-            "token-a",
-            "run-1",
-            [event(source_key="source-1")],
-            proxy_log_path,
-        )
-
-        try:
-            usage_webhook_server.queue_response(500)
-            usage_webhook_server.queue_response(500)
-            with patch.object(usage.webhook.time, "sleep"):
-                mitm_addon._handle_runner_usage_flush_signal(0, None)
-                wait_for_usage_flush_worker_to_stop()
-
-            assert usage_webhook_server.request_count == 2
-            failed_key = usage_webhook_server.requests[0].json_body()["events"][0]["idempotencyKey"]
-            assert_pending(
-                pending_path,
-                flows=0,
-                buffered=1,
-                reports=0,
-                flush_request_id="request-1",
-            )
-
-            usage_webhook_server.queue_response(204)
-            mitm_addon._handle_runner_usage_flush_signal(0, None)
-            wait_for_usage_flush_worker_to_stop()
-
-            assert usage_webhook_server.request_count == 3
-            retry_key = usage_webhook_server.requests[2].json_body()["events"][0]["idempotencyKey"]
-            assert retry_key == failed_key
-            assert_pending(
-                pending_path,
-                flows=0,
-                buffered=0,
-                reports=0,
-                flush_request_id="request-1",
-            )
-        finally:
-            wait_for_usage_flush_worker_to_stop()
-            usage.set_pending_path("")
-
-
-class TestTlsClienthello:
-    def test_unregistered_vm_ignored(self, registry_file, make_tls_data, mitm_ctx):
-        data = make_tls_data(client_ip="192.168.99.99")
-
-        with (
-            mitm_ctx(registry_path=str(registry_file), api_url="https://api.vm0.ai"),
-        ):
-            mitm_addon.tls_clienthello(data)
-
-        assert data.ignore_connection is True
-
-    def test_mitm_enabled_returns_early(self, registry_file, make_tls_data, mitm_ctx):
-        """When MITM is enabled, tls_clienthello should return without setting ignore_connection."""
-        data = make_tls_data(client_ip="10.200.0.1", sni="blocked.com")
-
-        with (
-            mitm_ctx(registry_path=str(registry_file), api_url="https://api.vm0.ai"),
-        ):
-            mitm_addon.tls_clienthello(data)
-
-        # MITM VM (10.200.0.1) should NOT set ignore_connection
-        assert data.ignore_connection is False
-
-    def test_registered_vm_allows_mitm(self, registry_file, make_tls_data, mitm_ctx):
-        """Registered VM does NOT set ignore_connection (allows MITM interception)."""
-        data = make_tls_data(client_ip="10.200.0.2", sni="anything.com")
-
-        with (
-            mitm_ctx(registry_path=str(registry_file), api_url="https://api.vm0.ai"),
-        ):
-            mitm_addon.tls_clienthello(data)
-
-        # All registered VMs use MITM — should NOT set ignore_connection
-        assert data.ignore_connection is False
-
-    def test_invalid_registered_vm_allows_mitm(self, tmp_path, make_tls_data, mitm_ctx):
-        registry_file = tmp_path / "registry.json"
-        registry_file.write_text(json.dumps({"vms": {"10.200.0.9": "broken"}, "updatedAt": 0}))
-        data = make_tls_data(client_ip="10.200.0.9", sni="anything.com")
-
-        with mitm_ctx(registry_path=str(registry_file), api_url="https://api.vm0.ai"):
-            mitm_addon.tls_clienthello(data)
-
-        assert data.ignore_connection is False
-
-    def test_registry_unavailable_does_not_ignore_connection(
-        self, registry_file, make_tls_data, mitm_ctx
-    ):
-        registry.load_registry(str(registry_file))
-        registry_file.unlink()
-        data = make_tls_data(client_ip="10.200.0.1", sni="anything.com")
-
-        with mitm_ctx(registry_path=str(registry_file), api_url="https://api.vm0.ai"):
-            mitm_addon.tls_clienthello(data)
-
-        assert data.ignore_connection is False
-
-
-class TestTcpStart:
-    def test_sets_metadata_for_registered_vm(self, registry_file, mitm_ctx, real_tcp_flow):
-        flow = real_tcp_flow(client_ip="10.200.0.1")
-
-        with (
-            mitm_ctx(registry_path=str(registry_file)),
-        ):
-            mitm_addon.tcp_start(flow)
-
-        assert flow.metadata["vm_run_id"] == "run-abc-123"
-        assert "vm_network_log_path" in flow.metadata
-        assert metadata_keys.TCP_START_MONOTONIC in flow.metadata
-
-    def test_skips_when_no_client_ip(self, registry_file, mitm_ctx, real_tcp_flow):
-        flow = real_tcp_flow()
-        flow.client_conn.peername = None
-
-        with (
-            mitm_ctx(registry_path=str(registry_file)),
-        ):
-            mitm_addon.tcp_start(flow)
-
-        assert "vm_run_id" not in flow.metadata
-
-    def test_skips_when_vm_not_registered(self, registry_file, mitm_ctx, real_tcp_flow):
-        flow = real_tcp_flow(client_ip="192.168.99.99")
-
-        with (
-            mitm_ctx(registry_path=str(registry_file)),
-        ):
-            mitm_addon.tcp_start(flow)
-
-        assert "vm_run_id" not in flow.metadata
-
-    def test_registry_unavailable_kills_flow(self, registry_file, mitm_ctx, real_tcp_flow):
-        registry.load_registry(str(registry_file))
-        registry_file.unlink()
-        flow = real_tcp_flow(client_ip="10.200.0.1")
-
-        with mitm_ctx(registry_path=str(registry_file)):
-            mitm_addon.tcp_start(flow)
-
-        assert flow.error is not None
-        assert flow.error.msg == Error.KILLED_MESSAGE
-        assert not flow.live
-        assert "vm_run_id" not in flow.metadata
-
-    def test_invalid_registered_vm_kills_flow(self, tmp_path, mitm_ctx, real_tcp_flow):
-        registry_file = tmp_path / "registry.json"
-        registry_file.write_text(json.dumps({"vms": {"10.200.0.9": {"runId": ""}}, "updatedAt": 0}))
-        flow = real_tcp_flow(client_ip="10.200.0.9")
-
-        with mitm_ctx(registry_path=str(registry_file)):
-            mitm_addon.tcp_start(flow)
-
-        assert flow.error is not None
-        assert flow.error.msg == Error.KILLED_MESSAGE
-        assert not flow.live
-        assert "vm_run_id" not in flow.metadata
-
-
-class TestTcpLog:
-    def test_logs_tcp_connection(self, registry_file, tmp_path, mitm_ctx, real_tcp_flow):
-        flow = real_tcp_flow(client_ip="10.200.0.1")
-        log_path = str(tmp_path / "network.jsonl")
-        flow.metadata["vm_run_id"] = "run-abc-123"
-        flow.metadata["vm_network_log_path"] = log_path
-        flow.metadata[metadata_keys.TCP_START_MONOTONIC] = time.monotonic() - 0.05
-
-        with mitm_ctx():
-            mitm_addon.tcp_end(flow)
-
-        entries = read_jsonl_entries_after_flush(Path(log_path))
-        assert len(entries) == 1
-        entry = entries[0]
-        assert entry["type"] == "tcp"
-        assert entry["host"] == "140.82.116.3"
-        assert entry["port"] == 22
-        assert entry["latency_ms"] > 0
-        assert entry["request_size"] == 5  # b"hello"
-        assert entry["response_size"] == 14  # b"SSH-2.0-babeld"
-        assert "error" not in entry
-        assert_utc_millisecond_timestamp(entry["timestamp"])
-
-    def test_logs_tcp_error(self, registry_file, tmp_path, mitm_ctx, real_tcp_flow):
-        flow = real_tcp_flow(client_ip="10.200.0.1")
-        log_path = str(tmp_path / "network.jsonl")
-        flow.metadata["vm_run_id"] = "run-abc-123"
-        flow.metadata["vm_network_log_path"] = log_path
-        flow.metadata[metadata_keys.TCP_START_MONOTONIC] = time.monotonic()
-        flow.error = Error("connection reset by peer")
-
-        with mitm_ctx():
-            mitm_addon.tcp_error(flow)
-
-        entry = read_jsonl_entries_after_flush(Path(log_path))[0]
-        assert entry["type"] == "tcp"
-        assert entry["error"] == "connection reset by peer"
-
-    def test_skips_when_no_run_id(self, tmp_path, mitm_ctx, real_tcp_flow):
-        flow = real_tcp_flow()
-        log_path = str(tmp_path / "network.jsonl")
-        flow.metadata["vm_network_log_path"] = log_path
-
-        with mitm_ctx():
-            mitm_addon.tcp_end(flow)
-
-        logging_utils.flush_log_path(log_path)
-        assert not Path(log_path).exists()
-
-    def test_handles_missing_server_addr(self, tmp_path, mitm_ctx, real_tcp_flow):
-        flow = real_tcp_flow()
-        log_path = str(tmp_path / "network.jsonl")
-        flow.metadata["vm_run_id"] = "run-abc-123"
-        flow.metadata["vm_network_log_path"] = log_path
-        flow.metadata[metadata_keys.TCP_START_MONOTONIC] = time.monotonic()
-        flow.server_conn = None
-
-        with mitm_ctx():
-            mitm_addon.tcp_end(flow)
-
-        [entry] = read_jsonl_entries_after_flush(Path(log_path))
-        assert entry["host"] == "unknown"
-        assert entry["port"] == 0
-
-    def test_handles_missing_start_time(self, tmp_path, mitm_ctx, real_tcp_flow):
-        flow = real_tcp_flow()
-        log_path = str(tmp_path / "network.jsonl")
-        flow.metadata["vm_run_id"] = "run-abc-123"
-        flow.metadata["vm_network_log_path"] = log_path
-
-        with mitm_ctx():
-            mitm_addon.tcp_end(flow)
-
-        [entry] = read_jsonl_entries_after_flush(Path(log_path))
-        assert entry["latency_ms"] == 0
-
-    def test_tcp_message_defers_registered_flow_drain(
-        self, tmp_path, monkeypatch, mitm_ctx, real_tcp_flow
-    ):
-        messages = [
-            tcp.TCPMessage(True, b"client-one"),
-            tcp.TCPMessage(False, b"server-one"),
-            tcp.TCPMessage(True, b"client-two"),
-        ]
-        flow = real_tcp_flow(messages=messages)
-        log_path = str(tmp_path / "network.jsonl")
-        flow.metadata["vm_run_id"] = "run-abc-123"
-        flow.metadata["vm_network_log_path"] = log_path
-        scheduled = _capture_tcp_drains(monkeypatch)
-
-        with mitm_ctx():
-            mitm_addon.tcp_message(flow)
-            mitm_addon.tcp_message(flow)
-
-        assert flow.messages == messages
-        assert len(scheduled) == 1
-
-        _run_scheduled_tcp_drains(scheduled)
-
-        assert flow.messages == []
-
-        with mitm_ctx():
-            mitm_addon.tcp_end(flow)
-
-        [entry] = read_jsonl_entries_after_flush(Path(log_path))
-        assert entry["request_size"] == len(b"client-one") + len(b"client-two")
-        assert entry["response_size"] == len(b"server-one")
-
-    def test_tcp_end_drains_pending_messages_before_deferred_callback(
-        self, tmp_path, monkeypatch, mitm_ctx, real_tcp_flow
-    ):
-        messages = [
-            tcp.TCPMessage(True, b"client"),
-            tcp.TCPMessage(False, b"server-response"),
-        ]
-        flow = real_tcp_flow(messages=messages)
-        log_path = str(tmp_path / "network.jsonl")
-        flow.metadata["vm_run_id"] = "run-abc-123"
-        flow.metadata["vm_network_log_path"] = log_path
-        scheduled = _capture_tcp_drains(monkeypatch)
-
-        with mitm_ctx():
-            mitm_addon.tcp_message(flow)
-            mitm_addon.tcp_end(flow)
-
-        [entry] = read_jsonl_entries_after_flush(Path(log_path))
-        assert entry["request_size"] == len(b"client")
-        assert entry["response_size"] == len(b"server-response")
-        assert flow.messages == []
-
-        _run_scheduled_tcp_drains(scheduled)
-        assert flow.messages == []
-
-    def test_tcp_message_reschedules_after_previous_drain(
-        self, tmp_path, monkeypatch, mitm_ctx, real_tcp_flow
-    ):
-        first_client = tcp.TCPMessage(True, b"first-client")
-        first_server = tcp.TCPMessage(False, b"first-server")
-        flow = real_tcp_flow(messages=[first_client, first_server])
-        log_path = str(tmp_path / "network.jsonl")
-        flow.metadata["vm_run_id"] = "run-abc-123"
-        flow.metadata["vm_network_log_path"] = log_path
-        scheduled = _capture_tcp_drains(monkeypatch)
-
-        with mitm_ctx():
-            mitm_addon.tcp_message(flow)
-
-        _run_scheduled_tcp_drains(scheduled)
-        assert flow.messages == []
-
-        second_client = tcp.TCPMessage(True, b"second-client")
-        second_server = tcp.TCPMessage(False, b"second-server")
-        flow.messages.extend([second_client, second_server])
-
-        with mitm_ctx():
-            mitm_addon.tcp_message(flow)
-
-        assert len(scheduled) == 1
-
-        _run_scheduled_tcp_drains(scheduled)
-
-        with mitm_ctx():
-            mitm_addon.tcp_end(flow)
-
-        [entry] = read_jsonl_entries_after_flush(Path(log_path))
-        assert entry["request_size"] == len(first_client.content) + len(second_client.content)
-        assert entry["response_size"] == len(first_server.content) + len(second_server.content)
-        assert flow.messages == []
-
-    def test_tcp_message_leaves_unregistered_flow_messages(
-        self, monkeypatch, mitm_ctx, real_tcp_flow
-    ):
-        messages = [tcp.TCPMessage(True, b"client")]
-        flow = real_tcp_flow(messages=messages)
-        scheduled = _capture_tcp_drains(monkeypatch)
-
-        with mitm_ctx():
-            mitm_addon.tcp_message(flow)
-
-        assert scheduled == []
-        assert flow.messages == messages
-
-    def test_tcp_size_counter_saturates_at_network_log_max(self, tmp_path, mitm_ctx, real_tcp_flow):
-        flow = real_tcp_flow(messages=[tcp.TCPMessage(True, b"overflow")])
-        log_path = str(tmp_path / "network.jsonl")
-        flow.metadata["vm_run_id"] = "run-abc-123"
-        flow.metadata["vm_network_log_path"] = log_path
-        flow.metadata[mitm_addon._TCP_REQUEST_SIZE] = mitm_addon._MAX_SAFE_NETWORK_LOG_SIZE - 1
-
-        with mitm_ctx():
-            mitm_addon.tcp_end(flow)
-
-        [entry] = read_jsonl_entries_after_flush(Path(log_path))
-        assert entry["request_size"] == mitm_addon._MAX_SAFE_NETWORK_LOG_SIZE
-        assert entry["response_size"] == 0
-        assert flow.messages == []
